refactor(watch): exceptions to errors.py, OverflowPolicy to types.py
- mirage/watch/errors.py holds QueueOverflowError + QueueClosed
  (matches the repo's per-package errors.py convention)
- OverflowPolicy StrEnum moves to mirage/types.py next to ChangeKind,
  the shared-leaf home for watch enums
- queue/base.py is now just the WatchQueue protocol + QueueFactory
- re-exports in queue/__init__ keep the public import paths stable
